refactor: Add `displayName` to a range of components
