+++ packages/ui/src/components/Display/Accordion/AccordionContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const AccordionContext = createContext(undefined)
