+++ packages/ui/src/components/Display/Accordion/index.ts
@@ -1,1 +1,2 @@
 export * from './Accordion'
+export * from './AccordionItem'
