+++ packages/ui/src/components/Display/List/ListIcon.tsx
@@ -72,3 +72,5 @@ export const ListIcon = forwardRef<
     </StyledListIcon>
   )
 })
+
+ListIcon.displayName = 'ListIcon'
