+++ packages/ui/src/components/Feedback/Tooltip/Tooltip.tsx
@@ -187,3 +187,5 @@ export const Tooltip = forwardRef<
     </StyledTooltip>
   )
 })
+
+Tooltip.displayName = 'Tooltip'
