+++ packages/ui/src/components/Input/Input/TextField.tsx
@@ -73,3 +73,5 @@ export const TextField = forwardRef<
     </InputWrapper>
   )
 })
+
+TextField.displayName = 'TextField'
