+++ packages/ui/src/components/Input/RadioGroup/RadioGroup.tsx
@@ -114,3 +114,5 @@ export const RadioGroup = forwardRef<ElementRef<typeof Group>, RadioGroupProps>(
     )
   }
 )
+
+RadioGroup.displayName = 'RadioGroup'
