+++ packages/ui/src/components/Input/Select/Select.tsx
@@ -253,3 +253,5 @@ export const Select = forwardRef<ElementRef<typeof StyledInput>, SelectProps>(
     )
   }
 )
+
+Select.displayName = 'Select'
