+++ packages/ui/src/components/Layout/Page.tsx
@@ -35,3 +35,5 @@ export const Page = forwardRef<ElementRef<typeof StyledPage>, PageProps>(
     )
   }
 )
+
+Page.displayName = 'Page'
