+++ packages/ui/src/components/Layout/ToggleMenuButton.tsx
@@ -39,3 +39,5 @@ export const ToggleMenuButton = forwardRef<
     </StyledToggleMenuButton>
   )
 })
+
+ToggleMenuButton.displayName = 'ToggleMenuButton'
