+++ packages/ui/src/components/Navigation/SideMenu/Menu/MenuItem.tsx
@@ -184,3 +184,5 @@ export const MenuItem = forwardRef<
     </>
   )
 })
+
+MenuItem.displayName = 'MenuItem'
