+++ packages/ui/src/components/Navigation/SideMenu/SideMenu.tsx
@@ -51,3 +51,5 @@ export const SideMenu = forwardRef<
     </StyledSideMenu>
   )
 })
+
+SideMenu.displayName = 'SideMenu'
