+++ packages/ui/src/components/Navigation/Tabs/Tab.tsx
@@ -93,3 +93,5 @@ export const Tab = forwardRef<ElementRef<typeof StyledTab>, TabProps>(
     )
   }
 )
+
+Tab.displayName = 'Tab'
