+++ packages/ui/src/components/Overlay/ContextMenu/ContextItem.tsx
@@ -37,3 +37,5 @@ export const ContextItem = forwardRef<
     </StyledContextItem>
   )
 })
+
+ContextItem.displayName = 'ContextItem'
