+++ packages/ui/src/components/Overlay/ContextMenu/ContextMenu.tsx
@@ -26,3 +26,5 @@ export const ContextMenu = forwardRef<
     </StyledBase>
   )
 })
+
+ContextMenu.displayName = 'ContextMenu'
