+++ packages/ui/src/components/Overlay/ContextMenu/ContextToggle.tsx
@@ -51,3 +51,5 @@ export const ContextToggle = forwardRef<
     </StyledContextToggle>
   )
 })
+
+ContextToggle.displayName = 'ContextToggle'
